requirements.txt
asgiref==3.8.1
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.9
Django==5.1.2
django-environ==0.11.2
djangorestframework==3.15.2
filelock==3.16.1
idna==2.10
packaging==24.1
pillow==11.0.0
pipenv==2024.2.0
platformdirs==4.3.6
psycopg2==2.9.10
python-decouple==3.8
pytz==2021.1
requests==2.25.1
setuptools==75.2.0
sqlparse==0.4.1
urllib3==1.26.4
virtualenv==20.27.0
djangorestframework-gis